fix rehash function, it now rehash all the node inside of it

+++ examples/array.c
@@ -3,7 +3,7 @@
 
 int main(void)
 {
-    Array(i32) ints = { 0 };
+    Array(i32) ints = {0};
     for (i32 i = 1; i <= 25; i++)
         array_append(&ints, i * 100);
 
@@ -13,7 +13,7 @@ int main(void)
     array_remove_at(&ints, 1);
     array_remove_at(&ints, ints.len - 1);
 
-    for (u32 i = 0; i < ints.len; i++)
+    for (u64 i = 0; i < ints.len; i++)
         printf("%d\n", ints.items[i]);
 
     array_free(&ints);
